design: moved data to repository
